Created new component Connection to fix some performance issues with the ConnectionsList
